Combine `login` and `logout` into `session` saga
